Move `thread_snapshot` to a dedicated header/source file pair.
