+++ CMakeLists.txt
@@ -8,6 +8,7 @@ src/profiling.cpp
 src/thread_pool.cpp
 src/thread_info.cpp
 src/thread_control_helper.cpp
+src/thread_snapshot.cpp
 )
 
 add_library(concore2full ${Sources})
