+++ src/thread_control_helper.cpp
@@ -1,14 +1,8 @@
 #include "thread_info.h"
 
 #include <concore2full/thread_control_helper.h>
-#include <concore2full/thread_reclaimer.h>
 
 #include <concore2full/detail/callcc.h>
-#include <concore2full/detail/thread_switch_helper.h>
-
-#include <mutex>
-#include <semaphore>
-#include <utility>
 
 namespace concore2full {
 
@@ -36,75 +30,6 @@ void thread_control_helper::check_for_thread_inversion() {
     });
     // The originating thread will continue this control flow.
   }
-}
-
-thread_snapshot::thread_snapshot() { original_thread_ = &detail::get_current_thread_info(); }
-
-void thread_snapshot::revert() {
-  // Are we on the same thread?
-  if (original_thread_ == &detail::get_current_thread_info()) {
-    // Good. No waiting needs to happen.
-  } else {
-    // Wait until we can start the switch.
-    // By the time waiting is over, we may be back on the desired thread, so don't do the switch in
-    // that case.
-    if (wait_for_switch_start())
-      perform_switch();
-  }
-}
-
-bool thread_snapshot::wait_for_switch_start() {
-  while (true) {
-    auto* cur_thread = &detail::get_current_thread_info();
-    if (cur_thread->switch_control_.request_switch_to(cur_thread, original_thread_)) {
-      // We started the switch process.
-      return true;
-    } else {
-      // We cannot switch at this moment.
-      // Check if another thread requested a switch from us.
-      thread_control_helper::check_for_thread_inversion();
-      // Now, this function may return on a different thread; check if we still need to switch.
-      if (&detail::get_current_thread_info() == original_thread_) {
-        return false;
-      }
-      // If we need to switch, introduce a small pause, so that we don't consume the CPU while
-      // spinning.
-      std::this_thread::yield();
-    }
-  }
-}
-
-void thread_snapshot::perform_switch() {
-  // Request a thread inversion. We might need to wait if our original thread is currently busy.
-  (void)detail::callcc([this](detail::continuation_t c) -> detail::continuation_t {
-    // Use the switch data from our thread.
-    auto& cur_thread = detail::get_current_thread_info();
-    auto* switch_data = &cur_thread.switch_data_;
-    switch_data->originator_start(c);
-    // TODO: fix race condition here; multiple threads may be trying to store at the same time
-    original_thread_->switch_control_.should_switch_with_.store(&cur_thread,
-                                                                std::memory_order_release);
-    // Note: After this line, the other thread can anytime switch to the `after_originator_`
-    // continuation, destryoing `this` pointer.
-
-    // If this thread is controlled by a thread pool which has a reclaimer registered, tell it to
-    // start reclaiming.
-    auto* reclaimer = original_thread_->thread_reclaimer_;
-    if (reclaimer)
-      reclaimer->start_reclaiming();
-
-    // Block until the original thread can perform the thread switch
-    cur_thread.switch_control_.waiting_semaphore_.acquire();
-
-    // The thread switch is complete.
-    // TODO: check for race conditions
-    cur_thread.switch_control_.switch_complete();
-
-    // Switch to the continuation provided by our original thread.
-    return cur_thread.switch_data_.originator_end();
-  });
-  // We resume here on the original thread.
-  assert(original_thread_ == &detail::get_current_thread_info());
 }
 
 } // namespace concore2full
