+++ src/thread_pool.cpp
@@ -2,6 +2,7 @@
 #include "concore2full/profiling.h"
 #include "concore2full/thread_control_helper.h"
 #include "concore2full/thread_reclaimer.h"
+#include "concore2full/thread_snapshot.h"
 
 #include <chrono>
